Use <pre> tag for commit message so that whitespace displays correctly
The contents of the commit message is inserted into the “#message” tag
directly.  If we use a <div> tag, then any whitespace in the commit
message will get collapsed down into a single space as HTML requires.
Since the CSS stylesheet calls for a monospaced font, it makes sense
to use a <pre> tag instead, so that whitespace will be preserved.